fix(#520): address second Copilot review round on station star PR
- Move tippy.css global import to _app.tsx (not valid in component files)
- Add type="button" to star and center-map buttons to prevent accidental
  form submissions if modal is ever used inside a <form>
- handleDragEnd now sets stationName:'Custom' when a waypoint is dragged,
  matching previous setWaypointCustomPosition behavior
- Validate localStorage JSON with Array.isArray before using as
  starredStations to guard against corrupted storage values
- Extract MapFlyTo into its own dynamic({ssr:false}) component to prevent
  react-leaflet's useMap from being imported at module scope in
  DeploymentMap, avoiding potential SSR/window-is-not-defined crashes
